refactor: extract protocol & metadata replay into log_segment submodule
Move protocol_and_metadata(), read_metadata(), and replay_for_metadata()
from log_segment.rs into log_segment/protocol_metadata_replay.rs. Methods
remain on impl LogSegment with no visibility or API changes.
